flask/requirements.txt
flask
pymongo
gunicorn
python-dotenv
pyjwt
flask-cors
google-cloud-storage
